requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
pymongo==4.5.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
Flask-Limiter==3.5.0
Flask-Validator==1.0.0
bcrypt==4.0.1
Pillow==10.0.0
gunicorn==21.2.0
python-magic==0.4.27
boto3==1.28.36
Flask-Cors==4.0.0
marshmallow==3.20.1
Flask-Mail==0.9.1 